fix unique key warning for Card.render
For lists of elements, React needs a unique `key` to determine what
gets re-rendered even when order changes. However, the `key` is
read-only and can be set only when creating the element. The way we
manipulated the children list here thus didn't work without a warning.
We now don't manipulate the children array ourselves anymore, so React
can take care of the keys itself.